# Core: Replace generator creation/iteration in CollectionState methods