@black-sliver 2025-01-31
When you posted the other result i already had the feeling that we can optimize this. I did however expect either a bigger or a smaller gain. Bigger because dicts are pretty fast, smaller because the optimizer could in theory do this (but seems like it doesn't).