@Mysteryem 2025-01-31
I think it is likely that the reason there isn't much visible change here is that the average world's performance is more affected by other areas, maybe simply that these functions are not used enough for this PR to make much of a difference on average.

I did not make a PR for these changes earlier because when I tried generations of 1 of every supported game template yaml, I did not observe a noticeable difference.

For the Wind Waker PR, where I've been looking at making the logic faster, I optimised the rules of 49 locations and added use of `has_all` to them. Doing so made the world generate slower.

Generating with 10 TWW yamls with every location enabled generates in 1.94s without this PR, but 1.49s with this PR, for a 23.3% reduction in generation duration.

Before my most recent local change that added the `has_all` usage in those 49 rules, those same 10 yamls would generate in 1.90s without this PR and 1.51s with this PR. So my optimisation made the multiworld generate 0.04s slower without this PR, but 0.02s faster with this PR.

Some of the games with higher numbers of text occurrences of `has_any(` and `has_all(` in `*.py`. This is not a particularly useful count because games can define a rule function once and then reference it repeatedly, which won't be counted here.
| Game | `has_any(` text occurrences | `has_all(` text occurrences | 10 template yamls before | 10 template yamls after | Reduction |
| --- | --- | --- | --- | --- | --- |
| The Wind Waker (local WIP) | 50 | 35 | 0.75s | 0.60s | -20.0% |
| Starcraft 2 | 101 | 38 | 4.6s | 3.9s | -15.2% |
| TUNIC | 45 | 12 | 0.214s | 0.210s | -1.9% |
| Yoshi's Island | 28 | 256 | 0.23s | 0.21s | - 8.7% |

![image](https://github.com/user-attachments/assets/b26db4f3-ff00-43d4-816a-3c09730f77fc) ![image](https://github.com/user-attachments/assets/4846f4d9-6725-4df0-a37c-4b8c8b2ac5d4)