@black-sliver 2025-02-01
The SC2 and WW gain is now higher than I would have expected :-D
Your tables are missing used python version and architecture, by the way.

Hotness of individual rules also depends on **which** locations have those rules and if the seed ends up in swap or not, so actual impact is always impossible to tell just by counting occurrences.

As a small note why "dicts are pretty fast", the hash used for dict access with string keys is cached in the str object and the dict implementation in python is amazingly fast.

I have not played around with this yet, but the code looks correct and I think we should merge this if someone finds the time to validate the improvement and compatibility.