@black-sliver 2025-02-02
Also did some testing, on py3.12 and py3.13, amd64, Linux, prog balancing on, spoiler and output off.
(Full spoiler with playthrough may also be impacted positively by this, but it probably adds more noise.)

For small games, improvements are easily reproducible by taking the avg of the better half of a larger number of identical gen runs. The numbers I personally looked at were close to noise, but reproducibly better.
E.g. 2 SoE + 1 Ts results in 1.2% - 1.6% faster gen on 3.13 and 3.12 even though SoE does not use any of the modified functions.

For huge multis, with the set of games I picked, the noise level is way beyond the improvement and it's impossible to run enough gens in a reasonable amount of time to de-noise it.